Refactoring * use Interfaces
